requirements.txt
# Mondo Poster Prompt - Python Dependencies

# Core dependencies (required for both scripts)
google-genai>=1.0.0

# Enhanced script dependencies (optional, only needed for generate_mondo_enhanced.py)
Pillow>=10.0.0  # For image-to-image and comparison features
